CLI: Move all the steps to index file
